feat: Rebuild icons
