--- packages/web/src/components/icons/Github.tsx
@@ -1,25 +1,0 @@
-import React from 'react';
-import { IconProps } from './types';
-
-export default function Github({
-  width = '19',
-  height = '19',
-  color = '#A5A5A5',
-  className,
-}: IconProps) {
-  return (
-    <svg
-      width={width}
-      height={height}
-      className={className}
-      viewBox='0 0 19 19'
-      fill='none'
-      xmlns='http://www.w3.org/2000/svg'
-    >
-      <path
-        d='M9.8075 0.678467C4.73273 0.678467 0.61731 4.80997 0.61731 9.90657C0.61731 13.9829 3.25063 17.4412 6.90418 18.6625C7.36375 18.7475 7.53149 18.4626 7.53149 18.2178C7.53149 17.9984 7.52345 17.4182 7.51885 16.6484C4.96251 17.2056 4.42252 15.411 4.42252 15.411C4.00547 14.3448 3.40228 14.061 3.40228 14.061C2.56702 13.4889 3.46433 13.5004 3.46433 13.5004C4.38691 13.5659 4.87175 14.4517 4.87175 14.4517C5.69208 15.8614 7.02367 15.4547 7.54758 15.218C7.6303 14.6217 7.86812 14.215 8.13123 13.9841C6.09075 13.7508 3.94457 12.9592 3.94457 9.42402C3.94457 8.41642 4.30304 7.59265 4.89128 6.94696C4.79592 6.71373 4.48112 5.77506 4.9809 4.50551C4.9809 4.50551 5.75297 4.25734 7.50851 5.45107C8.24152 5.24656 9.02738 5.14431 9.80979 5.14086C10.5899 5.14546 11.3769 5.24656 12.1111 5.45222C13.8655 4.25849 14.6364 4.50666 14.6364 4.50666C15.1373 5.77736 14.8225 6.71488 14.7283 6.94811C15.3177 7.5938 15.6727 8.41757 15.6727 9.42517C15.6727 12.9696 13.5242 13.7497 11.4769 13.9783C11.8066 14.2633 12.1007 14.8262 12.1007 15.6868C12.1007 16.9207 12.0892 17.9157 12.0892 18.2178C12.0892 18.4648 12.2547 18.7521 12.7211 18.6613C16.369 17.4389 19 13.9818 19 9.90657C19 4.80997 14.8846 0.678467 9.8075 0.678467Z'
-        fill={color}
-      />
-    </svg>
-  );
-}
