--- packages/web/src/components/icons/index.ts
@@ -1,4 +1,0 @@
-export { default as Grid } from './Grid';
-export { default as Lines } from './Lines';
-export { default as Github } from './Github';
-export { default as External } from './External';
