--- packages/web/src/components/icons/types.ts
@@ -1,7 +1,0 @@
-export type IconProps = {
-  width?: string;
-  height?: string;
-  className?: string;
-  color?: string;
-  onClick?: () => unknown;
-};
