+++ packages/web/src/components/ui/Vulnerability/Vulnerability.tsx
@@ -2,8 +2,8 @@ import React from 'react';
 import clsx from 'clsx';
 import TagBadge from '../TagBadge/TagBadge';
 import styles from './Vulnerability.module.scss';
-import { External } from '../../icons';
 import { Api } from '../../../services/apiClient';
+import { Icon } from '../Icon/Icon';
 
 export type Props = {
   vulnerability: Api.Vulnerability;
@@ -19,7 +19,7 @@ export default function Vulnerability({ vulnerability }: Props) {
           target='_blank'
           rel='noreferrer'
         >
-          <External />
+          <Icon kind={'external'} width={19} height={19} />
         </a>
       )}
